backend/requirements.txt

Django>=6.0
psycopg2-binary
djangorestframework
django-cors-headers
djangorestframework-simplejwt
requests>=2.31.0
python-dotenv>=1.0.0
